+++ scripts/effect-engine/effect-engine.js
@@ -2,6 +2,7 @@ import { componentRegistry } from "./component-registry.js";
 import { conditionComponent } from "./components/condition-component.js";
 import { modifierComponent } from "./components/modifier-component.js";
 import { persistentDamageComponent } from "./components/persistent-damage-component.js";
+import { resistanceComponent } from "./components/resistance-component.js";
 
 let initialized = false;
 
@@ -11,6 +12,7 @@ export function initializeEffectEngine() {
   componentRegistry.register(conditionComponent);
   componentRegistry.register(modifierComponent);
   componentRegistry.register(persistentDamageComponent);
+  componentRegistry.register(resistanceComponent);
 
   initialized = true;
 }